+++ R/addDriors.R
@@ -7,7 +7,9 @@
 #' @param priors data frame containing \code{stock}, \code{initial_state},
 #'        \code{initial_state_cv}, \code{terminal_state}, and
 #'        \code{terminal_state_cv}.
-#' @param stocks.combined whether to use the same priors for all stocks.
+#' @param same.priors whether to use the same priors for all stocks.
+#' @param stocks.combined old name for \code{same.prior}, provided for backward
+#'        compatibility. Use \code{same.prior} in new SOFIA scripts.
 #' @param shape_prior passed to \code{format_driors}.
 #' @param b_ref_type passed to \code{format_driors}.
 #' @param growth_rate_prior passed to \code{format_driors}.
@@ -15,11 +17,11 @@
 #' @param \dots additional arguments passed to \code{format_driors}.
 #'
 #' @details
-#' If \code{stocks.combined = TRUE} then the priors where \code{stock = All} are
+#' If \code{same.priors = TRUE} then the priors where \code{stock = All} are
 #' used for all stocks.
 #'
-#' If \code{stocks.combined = FALSE} then the priors are read from
-#' stock-specific rows in the \code{priors} data frame.
+#' If \code{same.priors = FALSE} then the priors are read from stock-specific
+#' rows in the \code{priors} data frame.
 #'
 #' @return
 #' A \code{tibble} like \code{stocks} but with the additional column
@@ -37,8 +39,8 @@
 #'
 #' @examples
 #' \dontrun{
-#' addDriors(stocks, priors, stocks.combined=TRUE)
-#' addDriors(stocks, priors, stocks.combined=FALSE)
+#' addDriors(stocks, priors, same.priors=TRUE)
+#' addDriors(stocks, priors, same.priors=FALSE)
 #' }
 #'
 #' @importFrom sraplus format_driors
@@ -46,15 +48,18 @@
 #'
 #' @export
 
-addDriors <- function(stocks, priors, stocks.combined, shape_prior=2,
-                      b_ref_type="k", growth_rate_prior=NA,
-                      growth_rate_prior_cv=0.2, ...)
+addDriors <- function(stocks, priors, same.priors=stocks.combined,
+                      stocks.combined, shape_prior=2, b_ref_type="k",
+                      growth_rate_prior=NA, growth_rate_prior_cv=0.2, ...)
 {
+  if(missing(same.priors) && missing(stocks.combined))
+    stop("argument 'same.priors' is missing")
+
   driors <- list()
   for(i in seq_len(nrow(stocks)))
   {
     ## p is the row number for the priors data frame
-    p <- if(stocks.combined) match("All", priors$stock)
+    p <- if(same.priors) match("All", priors$stock)
          else match(stocks$stock[i], priors$stock)
     driors[[i]] <- format_driors(
       taxa = stocks$taxa[i],